(#5) Evolui função buscar e buscar_caminho
